requirements.txt
certifi==2019.11.28
cffi==1.14.0
click==7.1.1
cryptography==2.8
decorator==4.4.2
future==0.18.2
importlib-metadata==1.6.0 ; python_version < '3.8'
importlib-resources==1.4.0 ; python_version < '3.7'
irc==18.0.0
jaraco.classes==3.1.0
jaraco.collections==3.0.0
jaraco.functools==3.0.0
jaraco.logging==3.0.0
jaraco.stream==3.0.0
jaraco.text==3.2.0
more-itertools==8.2.0
pycparser==2.20
python-telegram-bot==12.4.2
pytz==2019.3
six==1.14.0
tempora==3.0.0
toml==0.10.0
tornado==6.0.4
zipp==3.1.0 ; python_version < '3.8'
